auto_bot/pyproject.toml
[tool.poetry]
name = "auto-bot"
version = "0.1.0"
description = ""
authors = ["Klukvas <pavlenkoandrey56@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"
requests = "^2.31.0"
faker = "^22.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
